Add markers selection for aggregate series
